requirements.txt
pandas>=2.0
numpy>=1.26
matplotlib>=3.7
yfinance>=0.2.40
python-dateutil>=2.9
